INTERFACES back/fs/tree/Directory_Tree
t-Types are a OR of i-Type_props and i-Type. i-Types are a AND of i-Type_props and i-Types_meths.
It's enable to set only useful properties or a complete object as parameter
Cf. constructors

i-Types aggregate both i-Type_props and i-Types_meths, each one existing for easily overriding interfaces and constructing child t-Types in importing modules